backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Validation
pydantic[email]==2.10.0
pydantic-settings==2.6.0

# Task queue
celery==5.4.0
redis==5.2.0

# NLP / Embeddings
sentence-transformers==3.3.1
spacy==3.8.3
pdfminer.six==20231228
python-docx==1.1.2

# Scraping
jobspy==1.1.80
playwright==1.49.0
httpx==0.28.0
beautifulsoup4==4.12.3

# Utils
python-dotenv==1.0.1
tenacity==9.0.0
